# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "博邦 ERP Echo + GORM 模块化单体后台接口文档。后续新增、修改、删除接口时，必须同步更新 Swagger 注释、docs/API.md 和 test.http。",
        "title": "博邦 ERP 管理系统 API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "查询物料",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bb_erp_echo_internal_model.Material"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建物料主数据；必须选择当前账号部门下的在职员工。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "创建物料",
                "parameters": [
                    {
                        "description": "物料参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_material.createMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/bb_erp_echo_internal_model.Material"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_material.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bb_erp_echo_internal_model.Material": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "default_cost": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "operator_department_id": {
                    "type": "integer"
                },
                "operator_department_name": {
                    "type": "string"
                },
                "operator_employee_id": {
                    "type": "integer"
                },
                "operator_employee_name": {
                    "type": "string"
                },
                "operator_terminal_id": {
                    "type": "integer"
                },
                "operator_user_id": {
                    "type": "integer"
                },
                "operator_username": {
                    "type": "string"
                },
                "safety_stock": {
                    "type": "integer"
                },
                "spec": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_material.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 是错误码。",
                    "type": "string"
                },
                "message": {
                    "description": "Message 是前端可展示错误信息。",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID 是请求 ID，用于关联日志。",
                    "type": "string"
                }
            }
        },
        "internal_material.createMaterialRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "operator_employee_id"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "example": "生产物资"
                },
                "code": {
                    "type": "string",
                    "example": "M-001"
                },
                "default_cost": {
                    "type": "integer",
                    "example": 10000
                },
                "name": {
                    "type": "string",
                    "example": "铝材"
                },
                "operator_employee_id": {
                    "type": "integer",
                    "example": 1
                },
                "safety_stock": {
                    "type": "integer",
                    "example": 10
                },
                "spec": {
                    "type": "string",
                    "example": "标准"
                },
                "unit": {
                    "type": "string",
                    "example": "个"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 登录凭证，格式为：Bearer \u003ctoken\u003e。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}